+++ storage/src/db/schemas/collection.rs
@@ -54,7 +54,7 @@ RETURN IF $count IS NONE { 0 } ELSE IF $count.len() == 0 { 0 } ELSE { ($count[0]
 }"
         )
     )]
-    pub song_count: usize,
+    pub song_count: u64,
 }
 
 impl Collection {
